+++ src/awf/service/orphan_resources.py
@@ -9,7 +9,6 @@ import re
 import subprocess
 import time
 from collections.abc import Awaitable, Mapping
-from dataclasses import dataclass, field
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
 from typing import Any, Literal, Protocol
@@ -48,6 +47,33 @@ from awf.service.orphan_reaping import (
 )
 from awf.service.orphan_reaping import (
     build_orphan_compose_teardown as build_orphan_compose_teardown,
+)
+from awf.service.orphan_resource_types import (
+    RESOURCE_KINDS as RESOURCE_KINDS,
+)
+from awf.service.orphan_resource_types import (
+    ClassifiedResource as ClassifiedResource,
+)
+from awf.service.orphan_resource_types import (
+    CleanupReadiness as CleanupReadiness,
+)
+from awf.service.orphan_resource_types import (
+    DetectedResource as DetectedResource,
+)
+from awf.service.orphan_resource_types import (
+    DockerResourceCommand as DockerResourceCommand,
+)
+from awf.service.orphan_resource_types import (
+    OrphanResourceSummary as OrphanResourceSummary,
+)
+from awf.service.orphan_resource_types import (
+    ResourceKind as ResourceKind,
+)
+from awf.service.orphan_resource_types import (
+    ResourceScan as ResourceScan,
+)
+from awf.service.orphan_resource_types import (
+    WorkspaceIdView as WorkspaceIdView,
 )
 
 _log = get_logger(__name__)
@@ -111,10 +137,6 @@ class OrphanResourceWorktreeRemover(Protocol):
     ) -> Awaitable[WorkspaceGCWorktreeRemoveResult]: ...
 
 
-ResourceKind = Literal["container", "network", "volume", "worktree"]
-Classification = Literal["expected", "terminal", "missing", "unknown"]
-
-RESOURCE_KINDS: tuple[ResourceKind, ...] = ("container", "network", "volume", "worktree")
 ACTIVE_WORKSPACE_STATUSES = frozenset(
     {
         WorkspaceStatus.requested.value,
@@ -192,158 +214,6 @@ class AsyncCommandRunnerLike(Protocol):
         cwd: str | None = None,
         env: Mapping[str, str] | None = None,
     ) -> Any: ...
-
-
-@dataclass(frozen=True)
-class WorkspaceIdView:
-    """Snapshot of workspace ids partitioned by lifecycle."""
-
-    active_ids: frozenset[str]
-    terminal_ids: frozenset[str]
-    available: bool
-    retained_ids: frozenset[str] = frozenset()
-
-
-@dataclass(frozen=True)
-class DockerResourceCommand:
-    kind: ResourceKind
-    args: list[str]
-
-
-@dataclass(frozen=True)
-class DetectedResource:
-    kind: ResourceKind
-    workspace_id: str
-    compose_project: str | None = None
-    id: str | None = None
-    name: str | None = None
-    path: str | None = None
-    service: str | None = None
-    state: str | None = None
-    status_text: str | None = None
-    driver: str | None = None
-    scope: str | None = None
-
-
-@dataclass(frozen=True)
-class ClassifiedResource:
-    resource: DetectedResource
-    classification: Classification
-    reason: str
-
-    @property
-    def kind(self) -> ResourceKind:
-        return self.resource.kind
-
-    @property
-    def workspace_id(self) -> str:
-        return self.resource.workspace_id
-
-    @property
-    def compose_project(self) -> str | None:
-        return self.resource.compose_project
-
-    def to_dict(self) -> dict[str, object]:
-        payload: dict[str, object] = {
-            "kind": self.resource.kind,
-            "workspace_id": self.resource.workspace_id,
-            "classification": self.classification,
-            "reason": self.reason,
-        }
-        optional: dict[str, str | None] = {
-            "compose_project": self.resource.compose_project,
-            "id": self.resource.id,
-            "name": self.resource.name,
-            "path": self.resource.path,
-            "service": self.resource.service,
-            "state": self.resource.state,
-            "status": self.resource.status_text,
-            "driver": self.resource.driver,
-            "scope": self.resource.scope,
-        }
-        payload.update({key: value for key, value in optional.items() if value})
-        return payload
-
-
-@dataclass(frozen=True)
-class ResourceScan:
-    ok: bool
-    status: str
-    reason: str
-    resources: tuple[DetectedResource, ...] = ()
-    detail: str | None = None
-
-    def to_dict(self) -> dict[str, object]:
-        payload: dict[str, object] = {
-            "ok": self.ok,
-            "status": self.status,
-            "reason": self.reason,
-            "resource_count": len(self.resources),
-        }
-        if self.detail:
-            payload["detail"] = self.detail
-        return payload
-
-
-@dataclass(frozen=True)
-class CleanupReadiness:
-    ready: bool
-    status: str
-    reason: str
-    action: str
-    dry_run_only: bool = True
-
-    def to_dict(self) -> dict[str, object]:
-        return {
-            "ready": self.ready,
-            "status": self.status,
-            "reason": self.reason,
-            "action": self.action,
-            "dry_run_only": self.dry_run_only,
-        }
-
-
-@dataclass(frozen=True)
-class OrphanResourceSummary:
-    ok: bool
-    status: str
-    reason: str
-    resource_count: int
-    expected_count: int
-    orphan_count: int
-    unknown_count: int
-    counts_by_kind: dict[str, int]
-    orphan_counts_by_kind: dict[str, int]
-    expected_counts_by_kind: dict[str, int]
-    unknown_counts_by_kind: dict[str, int]
-    orphan_classification_counts: dict[str, int]
-    cleanup_readiness: CleanupReadiness
-    scanners: dict[str, dict[str, object]]
-    examples: tuple[dict[str, object], ...] = ()
-    detail: str | None = None
-    records: tuple[ClassifiedResource, ...] = field(default=(), repr=False)
-
-    def to_dict(self) -> dict[str, object]:
-        payload: dict[str, object] = {
-            "ok": self.ok,
-            "status": self.status,
-            "reason": self.reason,
-            "resource_count": self.resource_count,
-            "expected_count": self.expected_count,
-            "orphan_count": self.orphan_count,
-            "unknown_count": self.unknown_count,
-            "counts_by_kind": self.counts_by_kind,
-            "orphan_counts_by_kind": self.orphan_counts_by_kind,
-            "expected_counts_by_kind": self.expected_counts_by_kind,
-            "unknown_counts_by_kind": self.unknown_counts_by_kind,
-            "orphan_classification_counts": self.orphan_classification_counts,
-            "cleanup_readiness": self.cleanup_readiness.to_dict(),
-            "scanners": self.scanners,
-            "examples": list(self.examples),
-        }
-        if self.detail:
-            payload["detail"] = self.detail
-        return payload
 
 
 def docker_resource_commands() -> tuple[DockerResourceCommand, ...]:
