+++ tests/unit/node/test_git_manager.py
@@ -24,13 +24,6 @@ from awf.node.git_manager import (
 )
 
 
-@pytest.mark.unit
-def test_github_pull_head_ref_pattern_matches_expected() -> None:
-    pattern = git_manager._GITHUB_PULL_HEAD_REF
-    assert pattern.match("refs/pull/278/head")
-    assert pattern.match("refs/pull/0/head") is None
-
-
 def _git(args: list[str], cwd: Path) -> None:
     """Run a synchronous git command for fixture setup; fail loudly on error."""
     subprocess.run(
