+++ tests/unit/node/test_git_manager_ref_patterns.py
@@ -1,0 +1,11 @@
+"""Focused GitManager ref-pattern tests."""
+
+from __future__ import annotations
+
+import awf.node.git_manager as git_manager
+
+
+def test_github_pull_head_ref_pattern_matches_expected() -> None:
+    pattern = git_manager._GITHUB_PULL_HEAD_REF
+    assert pattern.match("refs/pull/278/head")
+    assert pattern.match("refs/pull/0/head") is None
